agent/pyproject.toml
[tool.poetry]
name = "aci"
version = "0.1.0"
description = "AI Agent project using Agno framework"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
agno = "*"
python-dotenv = "^1.0.0"
yfinance = "*"
openai = "^1.102.0"
sqlalchemy = "^2.0.43"
lancedb = "^0.24.3"
ddgs = "^9.5.5"
pypdf = "^5.1.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
gunicorn = "^23.0.0"
python-multipart = "^0.0.9"
python-telegram-bot = "^21.0"

[tool.poetry.group.prod.dependencies]
psycopg2-binary = "^2.9.9"
redis = "^5.0.0"
prometheus-client = "^0.20.0"
sentry-sdk = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
